Refactor supporting recursive object logic
- Revert ComplexObjectGenerator #263
- Add RecursiveObjectGuard

related #261
